Move distribute_tests and getallnodes to a separate module
