+++ reframe/frontend/autodetect.py
@@ -250,35 +250,3 @@ def detect_topology():
 
         if not found_devinfo:
             getlogger().debug(f'> device auto-detection is not supported')
-
-
-def getallnodes(state='all', jobs_cli_options=[]):
-    rt = runtime.runtime()
-    nodes = {}
-    for part in rt.system.partitions:
-        # This job will not be submitted, it's used only to filter
-        # the nodes based on the partition configuration
-        dummy_job = Job.create(part.scheduler,
-                               part.launcher_type(),
-                               name='placeholder-job',
-                               sched_access=part.access,
-                               sched_options=jobs_cli_options)
-
-        available_nodes = part.scheduler.allnodes()
-        available_nodes = part.scheduler.filternodes(dummy_job,
-                                                     available_nodes)
-        getlogger().debug(
-            f'Total available nodes for {part.name}: {len(available_nodes)}'
-        )
-
-        if state.casefold() != 'all':
-            available_nodes = {n for n in available_nodes
-                               if n.in_state(state)}
-            getlogger().debug(
-                f'[F] Selecting nodes in state {state!r}: '
-                f'available nodes now: {len(available_nodes)}'
-            )
-
-        nodes[part.fullname] = [n.name for n in available_nodes]
-
-    return nodes
